requirements.txt
pillow
roboflow